Exception Handling
Lots of errors lately relating to processing JSON coming from TMDb and Wikidata. Added more diagnostic information to help track down whether this is coming from a systemic change, or some kind of fault that hasn't been anticipated previously. Everything is already wrapped in several layers of try/except, so this is more of an effor to narrow it down without needing to add three more layers of try/except.