Error Handling
Added bit of code to check for when regen takes 00:00 - meaning ultimately that it failed. In such cases it waits 15min and issues an exception before trying again.  This is to stop, for example, retrying the same birthdate over and over again every few minutes for an hour while some remote service is offline.