Cache Skips
Send an email whenever the caching mechanism loops through all the birthdays and finds none to update. More of a means to check if the cache is spread out sufficiently, or whether there are long periods where no cache updates are required.  The idea generally is to space out the caching as much as possible over a week so that (1) it doesn't load the TMDb servers more than strictly necessary, (2) we don't make additional requests in the same week for data that doesn't really need to be refreshed and (3) we don't have periods where cached data has been deleted before it can be refreshed.