Cache Lookups
This update adds the mechanism that caches the "Lookup" endpoints. The main reason to do this is to speed up the Top1000 requests, but it may speed up others as well. The tradeoff is that these requests will now be stored in memory. Which isn't great. But as the server is restarted daily, it should work out ok, particularly as the top1000 values only change daily as well.
Would like to store a Brotli-encoded version so we can save memory and also save having to do the conversion. Next time perhaps.